+++ CHANGELOG.md
@@ -5,8 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+### Changed
 - [Patch] Allow Node types for SelectDropdown label prop
 - [Patch] Allow numbers for TextField value prop
+- [Patch] Allow Node types for Checkbox label prop
 
 ## 31.11.0 - 2018-08-01
 ### Added